getByte now returns int. Lots of debug. One bug fix.
Some external code was having issues converting the byte returned by
getByte to an int properly. I'm not sure this is the right answer but
we'll try it for now.

Trying to figure out what is going on with some subchunk oddities from
a different project.

ByteChunkSPI was always writing to array offset 0 on copyTo.