PairChunk*: Fix improper validation in copyTo(...)
This was checking bounds in a manner that it throwing IndexOutOfBounds
when the end index, which is exclusive, as if it were inclusive. In
particular this caused a copy of the full chunk to always fail.

See previous commits for unit test generator issues that cloaked this
problem from being detected.