{from,of,copy,give}(type) => {from,of,copy,give}Type(type) clarity
Using the library I find that I keep having to reference the source or
documentation to make sure it is doing what I want because of
overloading. This deprecates most overloaded varieties and replaces them
with methods containing the type in the name.

This also adds a getIntSize() method that throws if the size is greater
than an int can hold.