Performance: add getByte(int off) to SPI and direct from Chunk
It turns out that a great deal of time was being spent, especially on the FileChunksTest (next commit) converting from an int to a long and back to an int for BufferChunk with the associated validation. Chunk has always had a getByte(int) but it was just delegating directly to getByte(Long).

This commit changes Chunk and the SPI interfaces to allow the SPI itelf to decide if it wants to convert to long or not. In the common case, down to BufferChunk, this allows it to never convert to long and back.